feat: Add frequencySort function to array_9.ipynb
This commit adds a new function `frequencySort` to the `array_9.ipynb` file. The function uses the `Counter` class from the `collections` module to count the frequency of each element in the input list `nums`. It then sorts the list based on the frequency and the element value. The sorted list is returned as the result of the function.