feat: Add code to count the number of classes with more than 5 students
This commit adds code to count the number of classes with more than 5 students. It creates a DataFrame called "courses" with columns for student and class. The code groups the courses by class and counts the number of students in each class. It then filters the classes with 5 or more students and stores the result in a DataFrame called "result" with a column for class. This information can be used to analyze the distribution of class sizes.